optimisation: make lcd_encoder two bytes (int16_t)
lcd_encoder was int32_t (4 bytes) because of the menu code
when editing through the menus, the menus only accept
int16_t so it doesnt make sense to use int32_t.

Change in memory:
Flash: -892 bytes
SRAM: -2 bytes

+++ Firmware/menu.h
@@ -21,8 +21,8 @@ typedef struct
     //Variables used when editing values.
     const char* editLabel;
     void* editValue;
-    int32_t minEditValue;
-    int32_t maxEditValue;
+    int16_t minEditValue;
+    int16_t maxEditValue;
 } menu_data_edit_t;
 
 extern uint8_t menu_data[MENU_DATA_SIZE];
